requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
asttokens==2.4.1
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.3.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.1
cryptography==45.0.6
debugpy==1.8.1
decorator==5.1.1
Django==5.0.6
django-bootstrap3==24.3
djangorestframework==3.15.2
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.0.1
fastapi==0.117.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.2
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
multidict==6.6.4
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
passlib==1.7.4
Pillow==10.1.0
platformdirs==4.2.0
pluggy==1.6.0
prompt-toolkit==3.0.43
propcache==0.3.2
psutil==5.9.8
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.17.2
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.2
redis==5.0.1
requests==2.31.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.14
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.48.0
stripe==7.8.0
tornado==6.4
traitlets==5.14.1
twilio==8.10.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1

fastapi
uvicorn[standard]
gunicorn
